Refactors GTT settings
